+++ CHANGES.md
@@ -1,3 +1,18 @@
+## 0.0.2025XXXX (unreleased)
+
+### Added
+
+- Improve tests coverage (@mbarbin).
+- New flag `crs grep --summary` to display information as summary tables (@mbarbin).
+
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
 ## 0.0.20250414 (2025-04-14)
 
 This very early draft release is intended for publication to my custom opam-repository. It allows for initial experimentation and ensures that the release cycle and distribution process are functioning correctly.